+++ design-system/grid/examples/constellation/shared/skeleton-box.tsx
@@ -1,0 +1,18 @@
+import React, { FC } from 'react';
+
+import Box from '@atlaskit/ds-explorations/box';
+
+export const SkeletonBox: FC = ({ children }) => (
+  <Box
+    display="flex"
+    alignItems="center"
+    justifyContent="center"
+    backgroundColor="elevation.surface.sunken"
+    borderColor="color.border"
+    borderWidth="3px"
+    borderStyle="solid"
+    height="size.600"
+  >
+    {children}
+  </Box>
+);