+++ editor/editor-common/src/styles/shared/code-block.ts
@@ -23,8 +23,6 @@ import { borderRadius, fontSize, gridSize } from '@atlaskit/theme/constants';
 import { ThemeProps } from '@atlaskit/theme/types';
 import { token } from '@atlaskit/tokens';
 
-import { TaskDecisionSharedCssClassName } from './task-decision';
-
 export const CodeBlockSharedCssClassName = {
   CODEBLOCK_CONTAINER: 'code-block',
   CODEBLOCK_START: 'code-block--start',
@@ -149,11 +147,9 @@ export const codeBlockSharedStyles = (props: ThemeProps) => css`
 `;
 
 export const codeBlockInListSafariFix = css`
-  &:not(.${TaskDecisionSharedCssClassName.DECISION_CONTAINER}) {
-    ::before {
-      content: ' ';
-      line-height: ${akEditorLineHeight};
-    }
+  ::before {
+    content: ' ';
+    line-height: ${akEditorLineHeight};
   }
 
   > p:first-child,