+++ linking-platform/smart-card/src/view/LinkUrl/LinkWarningModal/hooks/use-link-warning-modal.tsx
@@ -1,10 +1,22 @@
 import { useState, MouseEvent } from 'react';
 import { normalizeUrl } from '@atlaskit/linking-common/url';
+import { useAnalyticsEvents } from '@atlaskit/analytics-next';
+import { ANALYTICS_CHANNEL } from '../../../../utils/analytics';
 
 export const useLinkWarningModal = () => {
   const [unsafeLinkText, setUnsafeLinkText] = useState<string | null>(null);
   const [url, setUrl] = useState<string | null>(null);
   const [isOpen, setIsOpen] = useState<boolean>(false);
+
+  const { createAnalyticsEvent } = useAnalyticsEvents();
+  const sendAnalyticsEvent = () => {
+    createAnalyticsEvent({
+      action: 'shown',
+      actionSubject: 'warningModal',
+      actionSubjectId: 'linkSafetyWarning',
+      eventType: 'operational',
+    }).fire(ANALYTICS_CHANNEL);
+  };
 
   /**
    * It checks and warns if a link text is a URL and different from an actual link destination
@@ -26,6 +38,7 @@ export const useLinkWarningModal = () => {
       setUnsafeLinkText(linkText);
       href && setUrl(href);
       setIsOpen(true);
+      sendAnalyticsEvent();
     }
   };