+++ linking-platform/smart-card/src/view/__tests__/linkUrl/linkUrl.test.tsx
@@ -1,6 +1,8 @@
 import React from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
 import LinkUrl from '../../LinkUrl';
+import { AnalyticsListener } from '@atlaskit/analytics-next';
+import { ANALYTICS_CHANNEL } from '../../../utils/analytics';
 
 describe('LinkUrl', () => {
   let LinkUrlTestId: string = 'link-with-safety';
@@ -44,5 +46,40 @@ describe('LinkUrl', () => {
     const isLinkUnsafe = !!screen.queryByText('Check this link');
 
     expect(isLinkUnsafe).toBe(expected);
+  });
+
+  it('should fire analytics event when link safety warning message is appeared', () => {
+    const onEvent = jest.fn();
+
+    const { getByTestId } = render(
+      <AnalyticsListener channel={ANALYTICS_CHANNEL} onEvent={onEvent}>
+        <LinkUrl href="https://some.url" checkSafety={true}>
+          https://another.url
+        </LinkUrl>
+      </AnalyticsListener>,
+    );
+
+    const LinkUrlView = getByTestId(LinkUrlTestId);
+    fireEvent.click(LinkUrlView);
+
+    expect(onEvent).toBeCalledWith(
+      expect.objectContaining({
+        hasFired: true,
+        payload: expect.objectContaining({
+          action: 'shown',
+          actionSubject: 'warningModal',
+          actionSubjectId: 'linkSafetyWarning',
+          eventType: 'operational',
+        }),
+        context: expect.arrayContaining([
+          {
+            componentName: 'linkUrl',
+            packageName: '@atlaskit/smart-card',
+            packageVersion: '999.9.9',
+          },
+        ]),
+      }),
+      ANALYTICS_CHANNEL,
+    );
   });
 });